refactor(cli): split upgrade into api/upgrade leaf shape
Reorganize the flat api/upgrade/upgrade.mjs into the fractal leaf shape:
- upgrade.mjs becomes a dispatcher + barrel (same exports: upgrade,
  refreshAgentDocs, UpgradeOptions), so api/index.mjs and the CLI consumer
  are unchanged.
- _adapter.mjs holds the shared machinery: version detection, agent-docs
  refresh, post-codemod hooks, and codemod selection/execution I/O.
- list/status/run leaves each project one response type (upgrade.list,
  upgrade.status {up_to_date|no_codemods|config_fixable}, upgrade.run).

Pure reorg: human output and --json envelopes are byte-identical, verified
by a differential harness across list/status/run/error cases (incl. the
config-repair apply on-disk result). Types stay central in types/upgrade.d.ts.
